+++ src/headless/plugins/bookmarks/tests/bookmarks.js
@@ -1,7 +1,7 @@
 import mock from '../../../tests/mock.js';
 import converse from '../../../dist/converse-headless.js';
 
-const { sizzle, stx, u } = converse.env;
+const { Strophe, sizzle, stx, u } = converse.env;
 
 describe('A bookmark', function () {
     it(
@@ -17,7 +17,7 @@ describe('A bookmark', function () {
 
             const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
             const sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('iq publish[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`iq publish[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(
@@ -27,9 +27,9 @@ describe('A bookmark', function () {
                     type="set"
                     xmlns="jabber:client">
                 <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                    <publish node="urn:xmpp:bookmarks:1">
+                    <publish node="${Strophe.NS.BOOKMARKS2}">
                         <item id="${muc.get('jid')}">
-                            <conference xmlns="urn:xmpp:bookmarks:1" autojoin="true" name="${settings.name}">
+                            <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="true" name="${settings.name}">
                                 <nick>${nick}</nick>
                                 <password>${settings.password}</password>
                             </conference>
@@ -65,7 +65,7 @@ describe('A bookmark', function () {
             id="${sent_stanza.getAttribute('id')}"/>`;
             _converse.api.connection.get()._dataRecv(mock.createRequest(_converse, stanza));
 
-            expect(muc.get('bookmarked')).toBeTruthy();
+            expect(muc.bookmark).toBeTruthy();
         }),
     );
 
@@ -82,7 +82,7 @@ describe('A bookmark', function () {
 
             const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
             let sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('iq publish[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`iq publish[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             const stanza = stx`<iq
@@ -101,7 +101,8 @@ describe('A bookmark', function () {
             }
 
             _converse.api.connection.get()._dataRecv(
-                mock.createRequest(_converse, 
+                mock.createRequest(
+                    _converse,
                     stx`<presence
                     xmlns="jabber:server"
                     from='${muc_jid}/${nick}'
@@ -123,7 +124,8 @@ describe('A bookmark', function () {
             await u.waitUntil(() => muc.get('nick') === newnick);
 
             _converse.api.connection.get()._dataRecv(
-                mock.createRequest(_converse, 
+                mock.createRequest(
+                    _converse,
                     stx`<presence
                     xmlns="jabber:server"
                     from='${muc_jid}/${newnick}'
@@ -140,7 +142,7 @@ describe('A bookmark', function () {
             );
 
             sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('iq publish[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`iq publish[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(
@@ -150,9 +152,9 @@ describe('A bookmark', function () {
                     type="set"
                     xmlns="jabber:client">
                 <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                    <publish node="urn:xmpp:bookmarks:1">
+                    <publish node="${Strophe.NS.BOOKMARKS2}">
                         <item id="${muc_jid}">
-                            <conference xmlns="urn:xmpp:bookmarks:1" name="${settings.name}" autojoin="true">
+                            <conference xmlns="${Strophe.NS.BOOKMARKS2}" name="${settings.name}" autojoin="true">
                                 <nick>${newnick}</nick>
                                 <password>${settings.password}</password>
                             </conference>
@@ -194,7 +196,7 @@ describe('A bookmark', function () {
 
                 const { bookmarks } = _converse.state;
 
-                let jid = 'theplay@conference.shakespeare.lit';
+                const jid = 'theplay@conference.shakespeare.lit';
                 const model = bookmarks.create({
                     jid,
                     autojoin: false,
@@ -246,7 +248,7 @@ describe('A bookmark', function () {
 
                 const { bookmarks } = _converse.state;
                 await u.waitUntil(() => bookmarks.length);
-                await u.waitUntil(() => muc.get('bookmarked'));
+                await u.waitUntil(() => muc.bookmark);
                 spyOn(bookmarks, 'sendBookmarkStanza').and.callThrough();
 
                 const sent_IQs = _converse.api.connection.get().IQ_stanzas;
@@ -268,9 +270,9 @@ describe('A bookmark', function () {
                         type="set"
                         xmlns="jabber:client">
                     <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                        <publish node="urn:xmpp:bookmarks:1">
+                        <publish node="${Strophe.NS.BOOKMARKS2}">
                             <item id="${muc_jid}">
-                                <conference xmlns="urn:xmpp:bookmarks:1" name="${settings.name}" autojoin="false">
+                                <conference xmlns="${Strophe.NS.BOOKMARKS2}" name="${settings.name}" autojoin="false">
                                     <nick>${nick}</nick>
                                 </conference>
                             </item>
@@ -320,15 +322,15 @@ describe('A bookmark', function () {
 
             const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
             let sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('publish[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(stx`
             <iq from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}" type="set" xmlns="jabber:client">
                 <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                    <publish node="urn:xmpp:bookmarks:1">
+                    <publish node="${Strophe.NS.BOOKMARKS2}">
                         <item id="${muc1_jid}">
-                            <conference xmlns="urn:xmpp:bookmarks:1" autojoin="true" name="Hamlet"/>
+                            <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="true" name="Hamlet"/>
                         </item>
                     </publish>
                     <publish-options>
@@ -363,16 +365,16 @@ describe('A bookmark', function () {
 
             sent_stanza = await u.waitUntil(() =>
                 IQ_stanzas.filter(
-                    (s) => sizzle('publish[node="urn:xmpp:bookmarks:1"] conference[name="Balcony"]', s).length,
+                    (s) => sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"] conference[name="Balcony"]`, s).length,
                 ).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(stx`
             <iq from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}" type="set" xmlns="jabber:client">
                 <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                    <publish node="urn:xmpp:bookmarks:1">
+                    <publish node="${Strophe.NS.BOOKMARKS2}">
                         <item id="${muc2_jid}">
-                            <conference xmlns="urn:xmpp:bookmarks:1" autojoin="true" name="Balcony">
+                            <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="true" name="Balcony">
                                 <nick>romeo</nick>
                             </conference>
                         </item>
@@ -414,16 +416,16 @@ describe('A bookmark', function () {
 
             sent_stanza = await u.waitUntil(() =>
                 IQ_stanzas.filter(
-                    (s) => sizzle('publish[node="urn:xmpp:bookmarks:1"] conference[name="Garden"]', s).length,
+                    (s) => sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"] conference[name="Garden"]`, s).length,
                 ).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(stx`
             <iq xmlns="jabber:client" type="set" from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}">
                 <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                    <publish node="urn:xmpp:bookmarks:1">
+                    <publish node="${Strophe.NS.BOOKMARKS2}">
                         <item id="${muc3_jid}">
-                            <conference xmlns="urn:xmpp:bookmarks:1" autojoin="false" name="Garden">
+                            <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="false" name="Garden">
                                 <nick>r0meo</nick>
                                 <password>secret</password>
                                 <extensions>
@@ -470,7 +472,7 @@ describe('A bookmark', function () {
 
             const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
             const sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle(`items[node="urn:xmpp:bookmarks:1"]`, s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`items[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             // Simulate server response with item-not-found error
@@ -480,7 +482,7 @@ describe('A bookmark', function () {
                         from="${sent_stanza.getAttribute('to')}"
                         to="${sent_stanza.getAttribute('from')}">
                     <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                        <items node="urn:xmpp:bookmarks:1"/>
+                        <items node="${Strophe.NS.BOOKMARKS2}"/>
                     </pubsub>
                     <error code="404" type="cancel">
                         <item-not-found xmlns="urn:ietf:params:xml:ns:xmpp-stanzas"/>
@@ -522,7 +524,7 @@ describe('A bookmark', function () {
 
             const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
             let sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('publish[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             // Server acknowledges successful storage
@@ -543,17 +545,257 @@ describe('A bookmark', function () {
 
             // Check that a retract stanza is sent as per XEP-0402
             sent_stanza = await u.waitUntil(() =>
-                IQ_stanzas.filter((s) => sizzle('retract[node="urn:xmpp:bookmarks:1"]', s).length).pop(),
+                IQ_stanzas.filter((s) => sizzle(`retract[node="${Strophe.NS.BOOKMARKS2}"]`, s).length).pop(),
             );
 
             expect(sent_stanza).toEqualStanza(stx`
                 <iq from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}" type="set" xmlns="jabber:client">
                     <pubsub xmlns="http://jabber.org/protocol/pubsub">
-                        <retract node="urn:xmpp:bookmarks:1" notify="true">
+                        <retract node="${Strophe.NS.BOOKMARKS2}" notify="true">
                             <item id="${muc_jid}"/>
                         </retract>
                     </pubsub>
                 </iq>`);
+        }),
+    );
+
+    it(
+        'can be pinned and sends out a stanza',
+        mock.initConverse(converse, ['connected', 'chatBoxesFetched'], {}, async function (_converse) {
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+
+            const bare_jid = _converse.session.get('bare_jid');
+            const muc_jid = 'theplay@conference.shakespeare.lit';
+            const { api, state } = _converse;
+
+            // First create a bookmark
+            state.bookmarks.create({
+                jid: muc_jid,
+                autojoin: true,
+                name: 'The Play',
+                nick: 'romeo',
+                extensions: [],
+            });
+
+            await mock.waitForMUCDiscoInfo(_converse, muc_jid);
+            await u.waitUntil(() => state.chatboxes.length === 1);
+
+            const IQ_stanzas = api.connection.get().IQ_stanzas;
+
+            // Now pin the bookmark
+            const bookmark = state.bookmarks.findWhere({ jid: muc_jid });
+            expect(bookmark).toBeTruthy();
+            await state.bookmarks.pinBookmark(bookmark);
+
+            const sent_stanza = await u.waitUntil(() =>
+                IQ_stanzas.filter(
+                    (s) =>
+                        sizzle(
+                            `publish[node="${Strophe.NS.BOOKMARKS2}"] conference[name="The Play"] extensions pinned`,
+                            s,
+                        ).length,
+                ).pop(),
+            );
+
+            expect(bookmark.get('pinned')).toBe(true);
+
+            const chatbox = state.chatboxes.get(muc_jid);
+            expect(chatbox.bookmark.get('pinned')).toBe(true);
+
+            expect(sent_stanza).toEqualStanza(stx`
+                <iq from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}" type="set" xmlns="jabber:client">
+                    <pubsub xmlns="http://jabber.org/protocol/pubsub">
+                        <publish node="${Strophe.NS.BOOKMARKS2}">
+                            <item id="${muc_jid}">
+                                <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="true" name="The Play">
+                                    <nick>romeo</nick>
+                                    <extensions>
+                                        <pinned xmlns="${Strophe.NS.BOOKMARKS_PINNING}"/>
+                                    </extensions>
+                                </conference>
+                            </item>
+                        </publish>
+                        <publish-options>
+                            <x type="submit" xmlns="jabber:x:data">
+                                <field type="hidden" var="FORM_TYPE">
+                                    <value>http://jabber.org/protocol/pubsub#publish-options</value>
+                                </field>
+                                <field var='pubsub#persist_items'>
+                                    <value>true</value>
+                                </field>
+                                <field var='pubsub#max_items'>
+                                    <value>max</value>
+                                </field>
+                                <field var='pubsub#send_last_published_item'>
+                                    <value>never</value>
+                                </field>
+                                <field var='pubsub#access_model'>
+                                    <value>whitelist</value>
+                                </field>
+                            </x>
+                        </publish-options>
+                    </pubsub>
+                </iq>`);
+        }),
+    );
+
+    it(
+        'can be unpinned and sends out a stanza',
+        mock.initConverse(converse, ['connected', 'chatBoxesFetched'], {}, async function (_converse) {
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+
+            const bare_jid = _converse.session.get('bare_jid');
+            const muc_jid = 'theplay@conference.shakespeare.lit';
+            const { api, state } = _converse;
+
+            // First create a pinned bookmark
+            const bookmark = state.bookmarks.create({
+                jid: muc_jid,
+                autojoin: true,
+                name: 'The Play',
+                nick: 'romeo',
+                extensions: [`<pinned xmlns="${Strophe.NS.BOOKMARKS_PINNING}"/>`],
+            });
+
+            await mock.waitForMUCDiscoInfo(_converse, muc_jid);
+            await u.waitUntil(() => state.chatboxes.length === 1);
+
+            const IQ_stanzas = api.connection.get().IQ_stanzas;
+
+            expect(bookmark.get('pinned')).toBe(true);
+            expect(state.chatboxes.get(muc_jid).bookmark.get('pinned')).toBe(true);
+
+            // Now unpin the bookmark
+            await state.bookmarks.unpinBookmark(bookmark);
+
+            const sent_stanza = await u.waitUntil(() =>
+                IQ_stanzas.filter((s) => {
+                    return sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"] conference[name="The Play"]`, s).length;
+                }).pop(),
+            );
+
+            expect(bookmark.get('pinned')).toBe(false);
+            expect(state.chatboxes.get(muc_jid).bookmark.get('pinned')).toBe(false);
+
+            expect(sent_stanza).toEqualStanza(stx`
+                <iq from="${bare_jid}" to="${bare_jid}" id="${sent_stanza.getAttribute('id')}" type="set" xmlns="jabber:client">
+                    <pubsub xmlns="http://jabber.org/protocol/pubsub">
+                        <publish node="${Strophe.NS.BOOKMARKS2}">
+                            <item id="${muc_jid}">
+                                <conference xmlns="${Strophe.NS.BOOKMARKS2}" autojoin="true" name="The Play">
+                                    <nick>romeo</nick>
+                                </conference>
+                            </item>
+                        </publish>
+                        <publish-options>
+                            <x type="submit" xmlns="jabber:x:data">
+                                <field type="hidden" var="FORM_TYPE">
+                                    <value>http://jabber.org/protocol/pubsub#publish-options</value>
+                                </field>
+                                <field var='pubsub#persist_items'>
+                                    <value>true</value>
+                                </field>
+                                <field var='pubsub#max_items'>
+                                    <value>max</value>
+                                </field>
+                                <field var='pubsub#send_last_published_item'>
+                                    <value>never</value>
+                                </field>
+                                <field var='pubsub#access_model'>
+                                    <value>whitelist</value>
+                                </field>
+                            </x>
+                        </publish-options>
+                    </pubsub>
+                </iq>`);
+        }),
+    );
+
+    it(
+        'normalizes extensions when publishing: drops malformed, dedups <pinned/>, preserves unknown',
+        mock.initConverse(converse, ['connected', 'chatBoxesFetched'], {}, async function (_converse) {
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+
+            const muc_jid = 'theplay@conference.shakespeare.lit';
+            const { state } = _converse;
+
+            // Already pinned, plus a malformed extension and an unrelated (unknown) one.
+            const bookmark = state.bookmarks.create({
+                jid: muc_jid,
+                autojoin: true,
+                name: 'The Play',
+                nick: 'romeo',
+                extensions: [
+                    '<not valid xml',
+                    '<note xmlns="urn:example:note"/>',
+                    `<pinned xmlns="${Strophe.NS.BOOKMARKS_PINNING}"/>`,
+                ],
+            });
+
+            await mock.waitForMUCDiscoInfo(_converse, muc_jid);
+            await u.waitUntil(() => state.chatboxes.length === 1);
+            expect(bookmark.get('pinned')).toBe(true);
+
+            const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
+
+            // Re-pinning must still publish (self-healing if the server diverged), and the
+            // published <extensions> must drop the malformed entry, keep exactly one
+            // <pinned/> (no duplicate) and preserve the unknown extension. (We don't await
+            // the publish, since the mock server never answers the IQ.)
+            state.bookmarks.pinBookmark(bookmark).catch(() => {});
+
+            const sent_stanza = await u.waitUntil(() =>
+                IQ_stanzas.filter(
+                    (s) => sizzle(`publish[node="${Strophe.NS.BOOKMARKS2}"] conference[name="The Play"]`, s).length
+                ).pop()
+            );
+
+            const children = Array.from(sizzle('extensions', sent_stanza)[0].children);
+            expect(children.length).toBe(2); // malformed dropped
+            expect(children.filter((c) => c.namespaceURI === Strophe.NS.BOOKMARKS_PINNING).length).toBe(1); // single <pinned/>
+            expect(children.filter((c) => c.namespaceURI === 'urn:example:note').length).toBe(1); // unknown preserved
+        }),
+    );
+
+    it(
+        'derives pinned-ness by element identity, not substring matching',
+        mock.initConverse(converse, ['connected', 'chatBoxesFetched'], {}, async function (_converse) {
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+
+            const { state } = _converse;
+            const ns = Strophe.NS.BOOKMARKS_PINNING;
+
+            // A lookalike the old substring check would have mis-flagged as pinned.
+            const lookalike = state.bookmarks.create({
+                jid: 'a@conference.example',
+                name: 'A',
+                autojoin: false,
+                extensions: [`<pinnedfoo xmlns="${ns}"/>`],
+            });
+            expect(lookalike.get('pinned')).toBe(false);
+
+            // A real <pinned/> with reordered/extra attributes, single quotes and
+            // an explicit close tag must still count as pinned.
+            const pinned = state.bookmarks.create({
+                jid: 'b@conference.example',
+                name: 'B',
+                autojoin: false,
+                extensions: [`<pinned foo='bar' xmlns="${ns}"></pinned>`],
+            });
+            expect(pinned.get('pinned')).toBe(true);
+
+            // A <pinned/> nested inside another extension must not count.
+            const nested = state.bookmarks.create({
+                jid: 'c@conference.example',
+                name: 'C',
+                autojoin: false,
+                extensions: [`<wrap xmlns="urn:example:wrap"><pinned xmlns="${ns}"/></wrap>`],
+            });
+            expect(nested.get('pinned')).toBe(false);
         }),
     );
 });